Set default value for corsAllowedOrigins in WebSecurity
The commit introduces a default value for the corsAllowedOrigins variable in the WebSecurity class. Prior to this commit, corsAllowedOrigins was not initialized, which could cause issues if it was not explicitly set later. Now, if the value is null or blank, it will automatically be set to "*", allowing requests from any origin.