Refactor package structure and add new user class and related files
The package structure of the project has been refactored to be more specific. The entities and repositories related to LeetCode problems have been moved from the general entity and repository packages to a new 'leetcode' subpackage. In addition, a new 'User' entity has been created. A corresponding repository, service, and controller for 'User' have also been added, including tests for the service and the controller. The project's 'README' and some test files were updated to reflect these changes.